Extract customer model logic into its own submodule; add remaining helpers to "utils".
