+++ perma_web/perma/models/__init__.py
@@ -1,18 +1,10 @@
 """
-Django models for the perma app, implemented in submodules (base.py, link.py, etc.).
+Django models for the perma app.
 
 Application code can keep using imports such as 'from perma.models import Link'.
 Two callables (get_default_archive_formats, get_empty_datetime_range) are also added here
 because older migrations resolve them on perma.models (migrations 0005 and 0046).
 """
-
-from .base import ( # noqa: F401
-    ACTIVE_SUBSCRIPTION_STATUSES,
-    FIELDS_REQUIRED_FROM_PERMA_PAYMENTS,
-    link_count_in_time_period,
-    most_active_org_in_time_period,
-    subscription_is_active
-)
 from .folder import Folder # noqa: F401
 from .internet_archive import ( # noqa: F401
     InternetArchiveFile,
@@ -37,4 +29,3 @@ from .user import( # noqa: F401
     LinkUserManager,
     UserOrganizationAffiliation
 )
-
