+++ perma_web/perma/models/customer.py
@@ -1,17 +1,14 @@
 import calendar
 from datetime import datetime
 from decimal import Decimal
-import logging
-
-import requests
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.db import models, transaction
-from django.db.models import Count
 from django.utils import timezone
 from django.utils.functional import cached_property
 from django.views.decorators.debug import sensitive_variables
-from taggit.models import CommonGenericTaggedItemBase, TaggedItemBase
+import requests
+import logging
 
 from perma.exceptions import InvalidTransmissionException, PermaPaymentsCommunicationException
 from perma.utils import (
@@ -23,6 +20,7 @@ from perma.utils import (
 )
 
 logger = logging.getLogger(__name__)
+
 
 ### CONSTANTS
 ACTIVE_SUBSCRIPTION_STATUSES = ['Current', 'Cancellation Requested']
@@ -47,38 +45,6 @@ CUSTOMER_TYPE_MAP = {
     'Registrar': 'Registrar'
 }
 
-
-### HELPERS ###
-
-# functions
-def link_count_in_time_period(links, start_time=None, end_time=None):
-    if start_time and end_time and (start_time > end_time):
-        raise ValueError("specified end time is earlier than specified start time")
-    elif start_time and end_time and (start_time == end_time):
-        links = links.filter(creation_timestamp=start_time)
-    else:
-        if start_time:
-            links = links.filter(creation_timestamp__gte=start_time)
-        if end_time:
-            links = links.filter(creation_timestamp__lte=end_time)
-    return links.count()
-
-def most_active_org_in_time_period(organizations, start_time=None, end_time=None):
-    if start_time and end_time and (start_time > end_time):
-        raise ValueError("specified end time is earlier than specified start time")
-    # unlike 'link_count_in_time_period', no special behavior required
-    # if start_time = end_time here. the end result is the same
-    else:
-        if start_time:
-            organizations = organizations.filter(links__creation_timestamp__gte=start_time)
-        if end_time:
-            organizations = organizations.filter(links__creation_timestamp__lte=end_time)
-        return organizations\
-            .annotate(num_links=Count('links'))\
-            .exclude(num_links=0)\
-            .order_by('-num_links')\
-            .first()
-
 def subscription_is_active(subscription):
     return subscription and (
         subscription['status'] in ACTIVE_SUBSCRIPTION_STATUSES or (
@@ -91,43 +57,6 @@ def subscription_is_active(subscription):
 def subscription_has_problem(subscription):
     return subscription and subscription['status'] in PROBLEM_SUBSCRIPTION_STATUSES
 
-
-# classes
-
-class DeletableManager(models.Manager):
-    """
-        Manager that excludes results where user_deleted=True by default.
-    """
-    def get_queryset(self):
-        # exclude deleted entries by default
-        return super(DeletableManager, self).get_queryset().filter(user_deleted=False)
-
-    def all_with_deleted(self):
-        return super(DeletableManager, self).get_queryset()
-
-
-class DeletableModel(models.Model):
-    """
-        Abstract base class that lets a model track deletion.
-    """
-    user_deleted = models.BooleanField(default=False, verbose_name="Deleted by user")
-    user_deleted_timestamp = models.DateTimeField(null=True, blank=True)
-
-    class Meta:
-        abstract = True
-
-    def safe_delete(self):
-        self.user_deleted = True
-        self.user_deleted_timestamp = timezone.now()
-
-
-# django-taggit assumes the model being tagged has an integer primary key.
-# per http://django-taggit.readthedocs.io/en/latest/custom_tagging.html,
-# tag "through" this class if your model has a string as primary key.
-# tags = TaggableManager(through=GenericStringTaggedItem)
-# (copied straight from their docs)
-class GenericStringTaggedItem(CommonGenericTaggedItemBase, TaggedItemBase):
-    object_id = models.CharField(max_length=50, db_index=True)
 
 
 class CustomerModel(models.Model):
@@ -617,4 +546,3 @@ class CustomerModel(models.Model):
         Must be implemented by children
         """
         raise NotImplementedError
-
