+++ perma_web/perma/tests/test_models.py
@@ -8,16 +8,11 @@ from django.utils import timezone
 from unittest.mock import patch, sentinel
 
 from perma.exceptions import PermaPaymentsCommunicationException, InvalidTransmissionException
-import perma.models
 from perma.models import (
-    ACTIVE_SUBSCRIPTION_STATUSES,
-    FIELDS_REQUIRED_FROM_PERMA_PAYMENTS,
-    Link, Organization, Registrar, Folder,
-    Sponsorship,
-    link_count_in_time_period,
-    most_active_org_in_time_period,
-    subscription_is_active
+    Link, Organization, Registrar, Folder, Sponsorship,
 )
+from perma.models.customer import ACTIVE_SUBSCRIPTION_STATUSES, FIELDS_REQUIRED_FROM_PERMA_PAYMENTS, subscription_is_active
+from perma.models.utils import link_count_in_time_period, most_active_org_in_time_period
 from perma.utils import pp_date_from_post, tz_datetime, first_day_of_next_month, today_next_month, today_next_year, years_ago_today
 
 from conftest import GENESIS
@@ -29,7 +24,7 @@ from waffle.testutils import override_switch
 # Related to Perma Payments
 #
 
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_none_and_no_network_call_if_nonpaying(post, noncustomers):
     # also verify that their value of in_trial is unchanged by the call
     for noncustomer in noncustomers:
@@ -40,7 +35,7 @@ def test_get_subscription_none_and_no_network_call_if_nonpaying(post, noncustome
         assert post.call_count == 0
 
 
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_raises_on_non_200(post, customers):
     post.return_value.ok = False
     for customer in customers:
@@ -50,8 +45,8 @@ def test_get_subscription_raises_on_non_200(post, customers):
         post.reset_mock()
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_verifies_transmission_valid(post, process, customers):
     post.return_value.status_code = 200
     post.return_value.json.return_value = sentinel.json
@@ -66,8 +61,8 @@ def test_get_subscription_verifies_transmission_valid(post, process, customers):
         process.reset_mock()
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_raises_if_unexpected_customer_pk(post, process, customers, spoof_pp_response_wrong_pk):
     post.return_value.status_code = 200
     for customer in customers:
@@ -78,8 +73,8 @@ def test_get_subscription_raises_if_unexpected_customer_pk(post, process, custom
         post.reset_mock()
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_raises_if_unexpected_registrar_type(post, process, customers, spoof_pp_response_wrong_type):
     post.return_value.status_code = 200
     for customer in customers:
@@ -90,8 +85,8 @@ def test_get_subscription_raises_if_unexpected_registrar_type(post, process, cus
         post.reset_mock()
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_no_subscription(post, process, customers, spoof_pp_response_no_subscription):
     post.return_value.status_code = 200
     for customer in customers:
@@ -110,8 +105,8 @@ def test_get_subscription_no_subscription(post, process, customers, spoof_pp_res
             post.reset_mock()
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_no_subscription_purchased_bonus(post, process, customers, spoof_pp_response_no_subscription_two_purchases):
     post.return_value.status_code = 200
     for customer in customers:
@@ -125,8 +120,8 @@ def test_get_subscription_no_subscription_purchased_bonus(post, process, custome
             post.reset_mock()
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_happy_path_sets_customer_trial_period_to_false(post, process, paying_registrar, paying_user, spoof_pp_response_subscription):
     post.return_value.status_code = 200
     for customer in [paying_registrar, paying_user]:
@@ -147,8 +142,8 @@ def test_get_subscription_happy_path_sets_customer_trial_period_to_false(post, p
         assert customer.cached_subscription_started
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_happy_path_no_change_pending(post, process, paying_registrar, paying_user, spoof_pp_response_subscription):
     post.return_value.status_code = 200
     for customer in [paying_registrar, paying_user]:
@@ -173,7 +168,7 @@ def test_get_subscription_happy_path_no_change_pending(post, process, paying_reg
 
 
 @override_switch('allow_cybersource_transactions', active=False)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_grandfathered_uses_cached_values(post, grandfathered_paying_user):
     subscription = grandfathered_paying_user.get_subscription()
 
@@ -190,8 +185,8 @@ def test_get_subscription_grandfathered_uses_cached_values(post, grandfathered_p
 
 
 @override_switch('allow_cybersource_transactions', active=False)
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_expired_grandfathered_unsets_flag_and_calls_perma_payments(
     post, process, expired_grandfathered_paying_user, spoof_pp_response_subscription,
 ):
@@ -217,8 +212,8 @@ def test_get_subscription_expired_grandfathered_unsets_flag_and_calls_perma_paym
     }
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_happy_path_with_pending_change(post, process, paying_user, spoof_pp_response_subscription_with_pending_change):
     post.return_value.status_code = 200
     customer = paying_user
@@ -247,8 +242,8 @@ def test_get_subscription_happy_path_with_pending_change(post, process, paying_u
         assert credited.call_count == 0
 
 
-@patch('perma.models.base.process_perma_payments_transmission', autospec=True)
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.process_perma_payments_transmission', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_get_subscription_null_effective_timestamp_treated_as_applied(post, process, paying_user, spoof_pp_response_subscription):
     # A null link_limit_effective_timestamp (e.g. a legacy Perma Payments row)
     # must not crash the usage-plan read on the None <= now comparison; it is
@@ -430,7 +425,7 @@ def test_user_link_creation_denied_when_frozen(get_links_remaining, nonpaying_us
     assert get_links_remaining.call_count == 0
 
 
-@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_allowed_checks_cached_if_pp_down(get_subscription, is_active, paying_user):
     get_subscription.side_effect = PermaPaymentsCommunicationException
@@ -464,8 +459,8 @@ def test_user_link_creation_denied_if_no_subscription_and_over_limit(get_subscri
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.base.subscription_is_active', autospec=True)
-@patch('perma.models.base.subscription_has_problem', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_has_problem', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_disallowed_if_subscription_inactive_and_over_limit(get_subscription, has_problem, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -480,8 +475,8 @@ def test_user_link_creation_disallowed_if_subscription_inactive_and_over_limit(g
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.base.subscription_is_active', autospec=True)
-@patch('perma.models.base.subscription_has_problem', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_has_problem', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_allowed_if_subscription_inactive_and_under_limit(get_subscription, has_problem, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -496,7 +491,7 @@ def test_user_link_creation_allowed_if_subscription_inactive_and_under_limit(get
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_allowed_if_subscription_active_and_under_limit(get_subscription, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -510,7 +505,7 @@ def test_user_link_creation_allowed_if_subscription_active_and_under_limit(get_s
 
 
 @patch('perma.models.LinkUser.links_remaining_in_period', autospec=True)
-@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
 @patch('perma.models.LinkUser.get_subscription', autospec=True)
 def test_user_link_creation_disallowed_if_subscription_active_and_under_limit(get_subscription, is_active, links_remaining_in_period, paying_user):
     get_subscription.return_value = sentinel.subscription
@@ -554,7 +549,7 @@ def test_registrar_link_creation_denied_when_frozen(get_subscription, nonpaying_
     assert get_subscription.call_count == 0
 
 
-@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
 @patch('perma.models.Registrar.get_subscription', autospec=True)
 def test_registrar_link_creation_allowed_checks_cached_if_pp_down(get_subscription, is_active, paying_registrar):
     get_subscription.side_effect = PermaPaymentsCommunicationException
@@ -573,8 +568,8 @@ def test_registrar_link_creation_disallowed_if_no_subscription(get_subscription,
     get_subscription.assert_called_once_with(paying_registrar)
 
 
-@patch('perma.models.base.subscription_is_active', autospec=True)
-@patch('perma.models.base.subscription_has_problem', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_has_problem', autospec=True)
 @patch('perma.models.Registrar.get_subscription', autospec=True)
 def test_registrar_link_creation_disallowed_if_subscription_inactive(get_subscription, has_problem, is_active, paying_registrar):
     get_subscription.return_value = sentinel.subscription
@@ -585,7 +580,7 @@ def test_registrar_link_creation_disallowed_if_subscription_inactive(get_subscri
     is_active.assert_called_once_with(sentinel.subscription)
 
 
-@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
 @patch('perma.models.Registrar.get_subscription', autospec=True)
 def test_registrar_link_creation_allowed_if_subscription_active(get_subscription, is_active, paying_registrar):
     get_subscription.return_value = sentinel.subscription
@@ -639,7 +634,7 @@ def test_annotate_tier_monthly_no_subscription_mid_month(customers):
         assert tier['next_payment'] == next_month
 
 
-@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
 def test_annotate_tier_monthly_no_subscription_last_of_month(is_active, customers):
     is_active.return_value = True
 
@@ -698,7 +693,7 @@ def test_annotate_tier_change_disallowed_with_inactive_annual_subscription(custo
             assert tier['type'] == 'unavailable'
 
 
-@patch('perma.models.base.subscription_is_active', autospec=True)
+@patch('perma.models.customer.subscription_is_active', autospec=True)
 def test_annotate_tier_change_disallowed_with_pending_downgrade(is_active, customers):
     is_active.return_value = True
 
@@ -1187,7 +1182,7 @@ def test_subscription_is_active_with_expired_canceled(expired_cancelled_subscrip
 # crediting users for one-time purchases
 #
 
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_credit_for_purchased_links_increments_for_all(post, paying_user, spoof_pp_response_no_subscription_two_purchases):
     post.return_value.ok = True
 
@@ -1198,7 +1193,7 @@ def test_credit_for_purchased_links_increments_for_all(post, paying_user, spoof_
     assert credited == 60
 
 
-@patch('perma.models.base.requests.post', autospec=True)
+@patch('perma.models.customer.requests.post', autospec=True)
 def test_credit_for_purchased_links_reverses_if_acknowledgment_fails(post, paying_user, spoof_pp_response_no_subscription_two_purchases):
     post.return_value.ok = False
 
@@ -1214,10 +1209,10 @@ def test_credit_for_purchased_links_reverses_if_acknowledgment_fails(post, payin
 # Bonus packages info
 #
 
-@patch.object(perma.models.base, 'datetime')
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.datetime')
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_get_bonus_packages(prep, mock_datetime, paying_user):
-    perma.models.base.datetime.utcnow.return_value = GENESIS
+    mock_datetime.utcnow.return_value = GENESIS
     prep.return_value.decode.return_value = sentinel.string
 
     packages = paying_user.get_bonus_packages()
