+++ perma_web/perma/tests/test_views_user_settings.py
@@ -221,7 +221,7 @@ def test_no_payment_message_for_unrecognized_params(client, user_without_subscri
     assert b'alert-block' not in response.content
 
 
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_subscribe_form_if_no_standing_subscription(prepped, client, user_without_subscription_or_purchase_history):
     user = user_without_subscription_or_purchase_history
     prepped.return_value = bytes(str(sentinel.prepped), 'utf-8')
@@ -346,7 +346,7 @@ def test_update_page_if_no_standing_subscription(client, user_without_subscripti
 
 
 @patch('perma.views.user_settings.prep_for_perma_payments', autospec=True)
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_update_page_if_standing_subscription(model_prepped, view_prepped, client, user_with_monthly_subscription):
     user = user_with_monthly_subscription
     model_prepped.return_value = bytes(str(sentinel.model_prepped), 'utf-8')
@@ -374,7 +374,7 @@ def test_update_page_if_standing_subscription(model_prepped, view_prepped, clien
 
 
 @patch('perma.views.user_settings.prep_for_perma_payments', autospec=True)
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_update_page_if_downgrade_scheduled(model_prepped, view_prepped, client, user_with_scheduled_downgrade):
     user = user_with_scheduled_downgrade
     model_prepped.return_value = bytes(str(sentinel.model_prepped), 'utf-8')
@@ -417,7 +417,7 @@ def test_update_page_if_subscription_on_hold(prepped, client, user_with_on_hold_
 
 # Subscriptions, Registrar Users
 
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_registrar_user_nonpaying_registrar(prepped, client, registrar_user_from_nonpaying_registrar):
     user = registrar_user_from_nonpaying_registrar
     prepped.return_value = bytes(str(sentinel.prepped), 'utf-8')
@@ -442,7 +442,7 @@ def test_registrar_user_nonpaying_registrar(prepped, client, registrar_user_from
     assert response.content.count(prepped.return_value) == individual_tier_count + bonus_package_count
 
 
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_paying_registrar_user_sees_both_subscribe_forms(prepped, client, registrar_user_from_paying_registrar_without_subscription):
     user = registrar_user_from_paying_registrar_without_subscription
     prepped.return_value = bytes(str(sentinel.prepped), 'utf-8')
@@ -466,7 +466,7 @@ def test_paying_registrar_user_sees_both_subscribe_forms(prepped, client, regist
     assert response.content.count(prepped.return_value) == tier_count + bonus_package_count
 
 
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_paying_registrar_user_sees_subscriptions_independently(prepped, client, registrar_user_from_registrar_with_monthly_subscription):
     user = registrar_user_from_registrar_with_monthly_subscription
     prepped.return_value = bytes(str(sentinel.prepped), 'utf-8')
@@ -496,7 +496,7 @@ def test_paying_registrar_user_sees_subscriptions_independently(prepped, client,
     assert response.content.count(b'<input type="hidden" name="account_type"') == 1
 
 
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_paying_registrar_user_sees_only_allowed_monthly_tier(prepped, client, registrar_user_from_paying_registrar_without_subscription):
     user = registrar_user_from_paying_registrar_without_subscription
     prepped.return_value = bytes(str(sentinel.prepped), 'utf-8')
@@ -514,7 +514,7 @@ def test_paying_registrar_user_sees_only_allowed_monthly_tier(prepped, client, r
     assert b'<span>Annual plan' not in response.content
 
 
-@patch('perma.models.base.prep_for_perma_payments', autospec=True)
+@patch('perma.models.customer.prep_for_perma_payments', autospec=True)
 def test_paying_registrar_user_sees_only_allowed_annual_tier(prepped, client, registrar_user_from_paying_registrar_without_subscription):
     user = registrar_user_from_paying_registrar_without_subscription
     prepped.return_value = bytes(str(sentinel.prepped), 'utf-8')
